+++ packages/editor/web/src/index.ts
@@ -18,3 +18,5 @@ export { RichContentEditorModal, RichContentEditor };
 export { getFontSize, getColor } from './RichContentEditor/utils/editorCommandsUtils';
 
 export { default as AddPluginMenu } from './RichContentEditor/Toolbars/SideToolbar/AddPluginMenu';
+
+export { SECTIONS } from './RichContentEditor/Toolbars/SideToolbar/utils';