+++ packages/ricos-modals/web/src/components/Popper.tsx
@@ -1,8 +1,9 @@
-import React, { useState } from 'react';
+import React, { useContext, useEffect, useState } from 'react';
 import { ClickOutside } from 'wix-rich-content-editor-common';
 import { usePopper } from 'react-popper';
 import type { ReactNode } from 'react';
 import type { ModalConfig } from 'ricos-types';
+import { ModalContext } from '../utils/ModalContext';
 
 interface Props {
   children: ReactNode;
@@ -13,6 +14,7 @@ interface Props {
 
 export const Popper = ({ children, modalConfig, closeModal, className }: Props) => {
   const [modalElement, setModalElement] = useState<HTMLDivElement | null>(null);
+  const { modalService } = useContext(ModalContext) || {};
 
   const { referenceElement, placement } = modalConfig.positioning || {};
   const { styles: popperStyles, attributes } = usePopper(referenceElement, modalElement, {
@@ -26,6 +28,20 @@ export const Popper = ({ children, modalConfig, closeModal, className }: Props) 
       },
     ],
   });
+
+  const closeOpenModalsWithSameRef = () => {
+    //Closes all open modals with the same reference element
+    const currentId = modalConfig.id;
+    modalService?.getOpenModals().find(({ id, positioning }) => {
+      const isSameRef = positioning?.referenceElement === referenceElement;
+      const shouldCloseModal = isSameRef && id !== currentId;
+      return shouldCloseModal && modalService?.closeModal(id);
+    });
+  };
+
+  useEffect(() => {
+    closeOpenModalsWithSameRef();
+  }, []);
 
   const onClickOutside = e => {
     !referenceElement.contains(e.target) && closeModal();